refactored TextNodeRepository
